Move custom audit log logic into the User class
